+++ backend/events/management/commands/consume.py
@@ -3,6 +3,7 @@ import time
 import structlog
 from django.core.management.base import BaseCommand
 
+from events import health
 from events.reactor import consume
 from events.reactor import resync
 
@@ -13,6 +14,7 @@ class Command(BaseCommand):
     help = "Consume events from the orchestrator broker"
 
     def handle(self, *args, **options):
+        health_service = health.HealthService()
         logger.info("resync local rep")
         # Init: resync all orchestrator assets
         resync()
@@ -21,7 +23,7 @@ class Command(BaseCommand):
         # Consume rabbitmq messages indefinitely
         while True:
             try:
-                consume()
+                consume(health_service)
             except Exception as e:
                 logger.exception("Error while consuming messages from the orchestrator RabbitMQ queue, will retry", e=e)
                 time.sleep(5)